check process state before clearing files directory

+++ pscload/src/main/java/fr/ans/psc/pscload/controller/ProcessController.java
@@ -20,7 +20,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.tomcat.util.http.fileupload.FileUtils;
+import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -72,7 +72,7 @@ public class ProcessController {
 
     @Autowired
     private EmailService emailService;
-    
+
     @Autowired
     private MessageProducer messageProducer;
 
@@ -211,13 +211,19 @@ public class ProcessController {
 
     @DeleteMapping(value = "/process/clear-files")
     public ResponseEntity<Void> clearFilesDirectory() {
-        try {
-            FileUtils.cleanDirectory(new File(filesDirectory));
-            log.info("cleaning directory {}", filesDirectory);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (IOException e) {
-            log.error("cleaning directory failed", e);
-            return new ResponseEntity<>(HttpStatus.OK);
+        LoadProcess process = registry.getCurrentProcess();
+
+        if (process == null) {
+            try {
+                FileUtils.cleanDirectory(new File(filesDirectory));
+                log.info("cleaning directory {}", filesDirectory);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } catch (IOException e) {
+                log.error("cleaning directory failed", e);
+                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
     }
 }
